# Add OTX optimize for visual prompting task