… CHANGELOG.md
… src/otx/algorithms/visual_prompting/configs/configuration.yaml
… src/otx/algorithms/visual_prompting/configs/sam_vit_b/config.yaml
… src/otx/algorithms/visual_prompting/configs/sam_vit_b/configuration.yaml
… tests/unit/algorithms/visual_prompting/adapters/pytorch_lightning/config/test_visual_prompting_config.py
… tests/unit/algorithms/visual_prompting/adapters/pytorch_lightning/datasets/pipelines/test_sam_transforms.py