### Summary

<!--
Resolves #111 and #222.
Depends on #1000 (for series of dependent commits).

This PR introduces this capability to make the project better in this and that.

- Added this feature
- Removed that feature
- Fixed the problem #1234
-->

This PR includes

- [x] enable `otx optimize` for ptq
- [x] remove unused modules
    - [x] transforms
    - [x] segment anything model
- [x] update config

#### PTQ results
  | torch | ov IR (fp32) | ov IR (ptq only image encoder)
-- | -- | -- | --
mDice | 0.9467 | 0.9442 | 0.9432

* Intel(R) Xeon(R) Gold 5320 CPU @ 2.20GHz

    throughput | f32 | u8
    -- | -- | --
    image   encoder | 1.84 fps | 1.97 fps
    decoder | 107.90 fps | -

* 11th Gen Intel(R) Core(TM) i9-11900 @ 2.50GHz

    throughput  | f32 | u8
    -- | -- | --
    image   encoder | 0.40 fps | 0.40 fps
    decoder | 84.00 fps | -

** PTQ for decoder (prompt encoder + mask decoder) is not supported because 
1. there are unknown issues when compiling compressed model
2. even if completed, and since decoder is very smaller than image encoder, it's not expected to improve performance significantly

### How to test

<!-- Describe the testing procedure for reviewers, if changes are
not fully covered by unit tests or manual testing can be complicated. -->

### Checklist

<!-- Put an 'x' in all the boxes that apply -->

- [x] I have added unit tests to cover my changes.​
- [ ] I have added integration tests to cover my changes.​
- [ ] I have added e2e tests for validation.
- [x] I have added the description of my changes into CHANGELOG in my target branch (e.g., [CHANGELOG](https://github.com/openvinotoolkit/training_extensions/blob/develop/CHANGELOG.md) in develop).​
- [ ] I have updated the documentation in my target branch accordingly (e.g., [documentation](https://github.com/openvinotoolkit/training_extensions/tree/develop/docs) in develop).
- [ ] I have [linked related issues](https://help.github.com/en/github/managing-your-work-on-github/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword).

### License

- [ ] I submit _my code changes_ under the same [Apache License](https://github.com/openvinotoolkit/training_extensions/blob/develop/LICENSE) that covers the project.
      Feel free to contact the maintainers if that's a concern.
- [ ] I have updated the license header for each file (see an example below).

```python
# Copyright (C) 2023 Intel Corporation
# SPDX-License-Identifier: Apache-2.0
```