@sungchul2 2023-07-13
> > > @sungchul2 do you have any accuracy results obtained with PTQ?
> > 
> > 
> > I added PTQ performance table in summary.
> 
> It seems like quantized model inference speed is almost the same as fp32. Did you try to run the inference on a machine with AVX-512 support?

Updated the table. As far as I know, those machines support AVX-512.