@sungchul2 2023-07-13
> LGTM, When the E2E tests will be added?

I'll do soon :)